Update PIRDS-v.0.1.md
@RobertLRead @ForrestErickson

I finished making most of the edits Lee pointed out, and also did a proofreading pass of the document. Let me know if you have any corrections, or if I changed the content.

The issues that are still outstanding because they would need a technical answer are Example 4, Example 6, Example 8, and Example 9 from the original errors list.